refactor(decisions): move voting eligibility from DB schema to service layer
Move VOTING_INELIGIBLE_STATUSES and isVotingEligible() out of
proposals.sql.ts into @op/common votingEligibility.ts. Business logic
belongs in the service layer, not the DB schema.
